% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncorrelation_analysis.R
\name{filter_factor_levels_by_missing}
\alias{filter_factor_levels_by_missing}
\title{Removes measurments with more than some percentage of missing values}
\usage{
filter_factor_levels_by_missing(pdata, config, percent = 60)
}
\arguments{
\item{pdata}{data.frame}

\item{config}{AnalysisConfiguration}

\item{percent}{percent of missing values}
}
\value{
data.frame
}
\description{
Removes measurments with more than some percentage of missing values
}
\examples{


bb <- prolfqua::sim_lfq_data_peptide_config()
config <- bb$config
config$parameter$min_nr_of_notNA  <- 20
data <- bb$data
data <- prolfqua::remove_large_QValues(data, config)
hc <- prolfqua::hierarchy_counts(data, config)
res <- filter_factor_levels_by_missing(data, config,percent = 80)

hc80 <-  prolfqua::hierarchy_counts(res, config)
res <- filter_factor_levels_by_missing(data, config,percent = 60)
prolfqua::hierarchy_counts(res, config)
hc60 <-  prolfqua::hierarchy_counts(res, config)
stopifnot(all(hc60 >= hc80)) # since 80\% missing is more stringent than 60\%
stopifnot(all(hc >= hc60))
prolfqua::summarize_hierarchy(res,config) |>
 dplyr::filter(!!rlang::sym(paste0(config$table$hierarchy_keys()[2],"_n")) > 1)

}
\keyword{internal}
