% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncorrelation_analysis.R
\name{impute_correlationBased}
\alias{impute_correlationBased}
\title{Imputate missing peptide intensities to maximize peptide correlationion}
\usage{
impute_correlationBased(x, config)
}
\arguments{
\item{x}{data}

\item{config}{configuration}
}
\description{
Assumes the peptide intensities are correlation assumption
}
\examples{



bb <- prolfqua::sim_lfq_data_peptide_config()

config <- bb$config$clone(deep=TRUE)
data <- bb$data
dataI <- impute_correlationBased(data, config)

}
\keyword{internal}
